Implement core-data module and integrate into the app module
