+++ app/src/main/kotlin/com/skydoves/pokedex/binding/ViewBinding.kt
@@ -32,7 +32,7 @@ import com.github.florent37.glidepalette.GlidePalette
 import com.google.android.material.card.MaterialCardView
 import com.skydoves.androidribbon.RibbonRecyclerView
 import com.skydoves.androidribbon.ribbonView
-import com.skydoves.pokedex.model.PokemonInfo
+import com.skydoves.pokedex.core.model.PokemonInfo
 import com.skydoves.pokedex.utils.PokemonTypeUtils
 import com.skydoves.pokedex.utils.SpacesItemDecoration
 import com.skydoves.progressview.ProgressView
