--- app/src/main/kotlin/com/skydoves/pokedex/di/RepositoryModule.kt
@@ -1,54 +1,0 @@
-/*
- * Designed and developed by 2020 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.skydoves.pokedex.di
-
-import com.skydoves.pokedex.network.PokedexClient
-import com.skydoves.pokedex.persistence.PokemonDao
-import com.skydoves.pokedex.persistence.PokemonInfoDao
-import com.skydoves.pokedex.repository.DetailRepository
-import com.skydoves.pokedex.repository.MainRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.android.scopes.ViewModelScoped
-import kotlinx.coroutines.CoroutineDispatcher
-
-@Module
-@InstallIn(ViewModelComponent::class)
-object RepositoryModule {
-
-  @Provides
-  @ViewModelScoped
-  fun provideMainRepository(
-    pokedexClient: PokedexClient,
-    pokemonDao: PokemonDao,
-    coroutineDispatcher: CoroutineDispatcher
-  ): MainRepository {
-    return MainRepository(pokedexClient, pokemonDao, coroutineDispatcher)
-  }
-
-  @Provides
-  @ViewModelScoped
-  fun provideDetailRepository(
-    pokedexClient: PokedexClient,
-    pokemonInfoDao: PokemonInfoDao,
-    coroutineDispatcher: CoroutineDispatcher
-  ): DetailRepository {
-    return DetailRepository(pokedexClient, pokemonInfoDao, coroutineDispatcher)
-  }
-}
