--- app/src/main/kotlin/com/skydoves/pokedex/model/Pokemon.kt
@@ -1,40 +1,0 @@
-/*
- * Designed and developed by 2020 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.skydoves.pokedex.model
-
-import android.os.Parcelable
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-import kotlinx.parcelize.Parcelize
-
-@Entity
-@Parcelize
-@JsonClass(generateAdapter = true)
-data class Pokemon(
-  var page: Int = 0,
-  @field:Json(name = "name") @PrimaryKey
-  val name: String,
-  @field:Json(name = "url") val url: String
-) : Parcelable {
-
-  fun getImageUrl(): String {
-    val index = url.split("/".toRegex()).dropLast(1).last()
-    return "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/$index.png"
-  }
-}
