--- app/src/main/kotlin/com/skydoves/pokedex/model/PokemonErrorResponse.kt
@@ -1,28 +1,0 @@
-/*
- * Designed and developed by 2020 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.skydoves.pokedex.model
-
-/**
- * A customized pokemon error response.
- *
- * @param code A network response code.
- * @param message A network error message.
- */
-data class PokemonErrorResponse(
-  val code: Int,
-  val message: String?
-)
