--- app/src/main/kotlin/com/skydoves/pokedex/persistence/PokemonDao.kt
@@ -1,36 +1,0 @@
-/*
- * Designed and developed by 2020 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.skydoves.pokedex.persistence
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.skydoves.pokedex.model.Pokemon
-
-@Dao
-interface PokemonDao {
-
-  @Insert(onConflict = OnConflictStrategy.REPLACE)
-  suspend fun insertPokemonList(pokemonList: List<Pokemon>)
-
-  @Query("SELECT * FROM Pokemon WHERE page = :page_")
-  suspend fun getPokemonList(page_: Int): List<Pokemon>
-
-  @Query("SELECT * FROM Pokemon WHERE page <= :page_")
-  suspend fun getAllPokemonList(page_: Int): List<Pokemon>
-}
