+++ app/src/main/res/layout/item_pokemon.xml
@@ -22,8 +22,8 @@
 
     <variable
       name="pokemon"
-      type="com.skydoves.pokedex.model.Pokemon" />
-
+      type="com.skydoves.pokedex.core.model.Pokemon" />
+    
   </data>
 
   <com.skydoves.transformationlayout.TransformationLayout
