--- app/src/test/kotlin/com/skydoves/pokedex/MainCoroutinesRule.kt
@@ -1,44 +1,0 @@
-/*
- * Designed and developed by 2020 skydoves (Jaewoong Eum)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress("SpellCheckingInspection")
-
-package com.skydoves.pokedex
-
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.test.TestDispatcher
-import kotlinx.coroutines.test.TestScope
-import kotlinx.coroutines.test.UnconfinedTestDispatcher
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
-import org.junit.rules.TestWatcher
-import org.junit.runner.Description
-
-class MainCoroutinesRule(
-  val testDispatcher: TestDispatcher = UnconfinedTestDispatcher()
-) : TestWatcher() {
-
-  val testScope = TestScope(testDispatcher)
-
-  override fun starting(description: Description) {
-    Dispatchers.setMain(testDispatcher)
-  }
-
-  override fun finished(description: Description) {
-    super.finished(description)
-    Dispatchers.resetMain()
-  }
-}
