+++ build.gradle.kts
@@ -41,6 +41,7 @@ subprojects {
     kotlinOptions.jvmTarget = JavaVersion.VERSION_1_8.toString()
     kotlinOptions.freeCompilerArgs += listOf(
       "-Xopt-in=kotlinx.coroutines.ExperimentalCoroutinesApi",
+      "-Xopt-in=kotlin.time.ExperimentalTime",
     )
   }
