+++ core-data/build.gradle.kts
@@ -1,0 +1,42 @@
+import com.skydoves.pokedex.Configuration
+
+plugins {
+  id(libs.plugins.android.library.get().pluginId)
+  id(libs.plugins.kotlin.android.get().pluginId)
+  id(libs.plugins.kotlin.kapt.get().pluginId)
+}
+
+android {
+  compileSdk = Configuration.compileSdk
+
+  defaultConfig {
+    minSdk = Configuration.minSdk
+    targetSdk = Configuration.targetSdk
+  }
+}
+
+dependencies {
+  api(project(":core-model"))
+  implementation(project(":core-network"))
+  implementation(project(":core-database"))
+  testImplementation(project(":core-test"))
+
+  // coroutines
+  implementation(libs.coroutines)
+  testImplementation(libs.coroutines)
+  testImplementation(libs.coroutines.test)
+
+  // network
+  implementation(libs.sandwich)
+
+  // di
+  implementation(libs.hilt.android)
+  kapt(libs.hilt.compiler)
+
+  // unit test
+  testImplementation(libs.junit)
+  testImplementation(libs.turbine)
+  testImplementation(libs.androidx.test.core)
+  testImplementation(libs.mockito.kotlin)
+  testImplementation(libs.mockito.inline)
+}
