+++ core-data/src/main/kotlin/com/skydoves/core/data/repository/MainRepositoryImpl.kt
@@ -1,0 +1,70 @@
+/*
+ * Designed and developed by 2020 skydoves (Jaewoong Eum)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.skydoves.core.data.repository
+
+import androidx.annotation.WorkerThread
+import com.skydoves.pokedex.core.database.PokemonDao
+import com.skydoves.pokedex.core.database.entitiy.mapper.asDomain
+import com.skydoves.pokedex.core.database.entitiy.mapper.asEntity
+import com.skydoves.pokedex.core.model.Pokemon
+import com.skydoves.pokedex.core.network.Dispatcher
+import com.skydoves.pokedex.core.network.PokedexAppDispatchers
+import com.skydoves.pokedex.core.network.service.PokedexClient
+import com.skydoves.sandwich.ApiResponse
+import com.skydoves.sandwich.message
+import com.skydoves.sandwich.onFailure
+import com.skydoves.sandwich.suspendOnSuccess
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.onCompletion
+import kotlinx.coroutines.flow.onStart
+import javax.inject.Inject
+
+class MainRepositoryImpl @Inject constructor(
+  private val pokedexClient: PokedexClient,
+  private val pokemonDao: PokemonDao,
+  @Dispatcher(PokedexAppDispatchers.IO) private val ioDispatcher: CoroutineDispatcher
+) : MainRepository {
+
+  @WorkerThread
+  override fun fetchPokemonList(
+    page: Int,
+    onStart: () -> Unit,
+    onComplete: () -> Unit,
+    onError: (String?) -> Unit
+  ) = flow {
+    var pokemons = pokemonDao.getPokemonList(page).asDomain()
+    if (pokemons.isEmpty()) {
+      /**
+       * fetches a list of [Pokemon] from the network and getting [ApiResponse] asynchronously.
+       * @see [suspendOnSuccess](https://github.com/skydoves/sandwich#apiresponse-extensions-for-coroutines)
+       */
+      val response = pokedexClient.fetchPokemonList(page = page)
+      response.suspendOnSuccess {
+        pokemons = data.results
+        pokemons.forEach { pokemon -> pokemon.page = page }
+        pokemonDao.insertPokemonList(pokemons.asEntity())
+        emit(pokemonDao.getAllPokemonList(page).asDomain())
+      }.onFailure { // handles the all error cases from the API request fails.
+        onError(message())
+      }
+    } else {
+      emit(pokemonDao.getAllPokemonList(page).asDomain())
+    }
+  }.onStart { onStart() }.onCompletion { onComplete() }.flowOn(ioDispatcher)
+}
