+++ core-database/src/main/kotlin/com/skydoves/pokedex/core/database/PokedexDatabase.kt
@@ -24,7 +24,7 @@ import com.skydoves.pokedex.core.database.entitiy.PokemonInfoEntity
 
 @Database(
   entities = [PokemonEntity::class, PokemonInfoEntity::class],
-  version = 1,
+  version = 2,
   exportSchema = true
 )
 @TypeConverters(value = [TypeResponseConverter::class])
