+++ core-network/src/main/kotlin/com/skydoves/pokedex/core/network/model/mapper/ErrorResponseMapper.kt
@@ -1,0 +1,40 @@
+/*
+ * Designed and developed by 2020 skydoves (Jaewoong Eum)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.skydoves.pokedex.core.network.model.mapper
+
+import com.skydoves.pokedex.core.network.model.PokemonErrorResponse
+import com.skydoves.sandwich.ApiErrorModelMapper
+import com.skydoves.sandwich.ApiResponse
+import com.skydoves.sandwich.message
+
+/**
+ * A mapper for mapping [ApiResponse.Failure.Error] response as custom [PokemonErrorResponse] instance.
+ *
+ * @see [ApiErrorModelMapper](https://github.com/skydoves/sandwich#apierrormodelmapper)
+ */
+object ErrorResponseMapper : ApiErrorModelMapper<PokemonErrorResponse> {
+
+  /**
+   * maps the [ApiResponse.Failure.Error] to the [PokemonErrorResponse] using the mapper.
+   *
+   * @param apiErrorResponse The [ApiResponse.Failure.Error] error response from the network request.
+   * @return A customized [PokemonErrorResponse] error response.
+   */
+  override fun map(apiErrorResponse: ApiResponse.Failure.Error<*>): PokemonErrorResponse {
+    return PokemonErrorResponse(apiErrorResponse.statusCode.code, apiErrorResponse.message())
+  }
+}
